PYCBC-463: Make TxIOEvent API compatible with Tornado
Motivation
----------

TxCouchbase can be used in a variety of environments, including within
a Tornado application, which converts twisted objects into native
Tornado objects.
When a connection is lost, Tornado tries to execute a callback to
TxCouchbase to inform it. Unfortunately the API it uses does not match
the API of the TxIOEvent, this means that the program will crash due to
an AttributeError.

Modifications
-------------

Added two new methods to TxIOEvent, readConnectionLost and
writeConnectionLost.
Both of these methods redirect to the existing method connectionLost,
this ensures that the API does not break for any existing users.

Now TxCouchbase will not crash when a connection fails when running in
Tornado.

Change-Id: Ia9d22731d1531359d2dbf5b73f90633c1b762802
Reviewed-on: http://review.couchbase.org/87849
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>